+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.9.12] 2021
+### Improvements
+- Arch
+    - repositories/AUR search time (=~ -70%)
 ### Fixes
 - Arch
     - crashing when information of a given package is not available